+++ src-docs/src/views/datagrid/datagrid_example.js
@@ -29,6 +29,7 @@ import {
   DataGridToolbarVisibilityOptions,
   DataGridColumnVisibility,
   DataGridPopoverContent,
+  DataGridControlColumn,
 } from './props';
 
 const gridSnippet = `
@@ -46,6 +47,22 @@ const gridSnippet = `
       visibleColumns: ['A', 'C'],
       setVisibleColumns: () => {},
     }}
+    leadingControlColumns={[
+      {
+        id: 'selection',
+        width: 31,
+        headerCellRender: () => <span>Select a Row</span>,
+        rowCellRender: () => <div><EuiCheckbox ... /></div>,
+      },
+    ]}
+    trailingControlColumns={[
+      {
+        id: 'actions',
+        width: 40,
+        headerCellRender: () => null,
+        rowCellRender: MyGridActionsComponent,
+      },
+    ]}
     // Optional. Customize the content inside the cell. The current example outputs the row and column position.
     // Often used in combination with useEffect() to dynamically change the render.
     renderCellValue={({ rowIndex, columnId }) =>
@@ -158,6 +175,16 @@ const gridConcepts = [
         An array of <EuiCode>EuiDataGridColumnVisibility</EuiCode> objects.
         Defines which columns are visible in the grid and the order they are
         displayed.
+      </span>
+    ),
+  },
+  {
+    title: 'leading and trailing controlColumns',
+    description: (
+      <span>
+        An array of <EuiCode>EuiDataGridControlColumn</EuiCode> objects. Used to
+        define ancillary columns on the left side of the data grid. Useful for
+        adding items like checkboxes and buttons.
       </span>
     ),
   },
@@ -299,6 +326,13 @@ export const DataGridExample = {
               can be controlled by the engineer, but augmented by user
               preference depending upon the features you enable.
             </li>
+            <li>
+              <Link to="/tabular-content/data-grid-control-collumns/">
+                Control columns
+              </Link>{' '}
+              allow you to add repeatable actions and controls like checkboxes
+              or buttons to your grid.
+            </li>
           </ul>
         </Fragment>
       ),
@@ -307,6 +341,7 @@ export const DataGridExample = {
         EuiDataGrid,
         EuiDataGridColumn: DataGridColumn,
         EuiDataGridColumnVisibility: DataGridColumnVisibility,
+        EuiDataGridControlColumn: DataGridControlColumn,
         EuiDataGridInMemory: DataGridInMemory,
         EuiDataGridPagination: DataGridPagination,
         EuiDataGridSorting: DataGridSorting,